+++ packages/impromat/src/pages/PrivacyPolicyPage.tsx
@@ -9,6 +9,8 @@ import {
   IonToolbar,
   useIonRouter,
 } from "@ionic/react";
+import ReactMarkdown from "react-markdown";
+import { datenschutzMarkdown } from "../markdown/datenschutz-markdown.md";
 
 export const PrivacyPolicyPage: React.FC = () => {
   const router = useIonRouter();
@@ -36,12 +38,27 @@ export const PrivacyPolicyPage: React.FC = () => {
           limited list of beta users we will complete the privacy policy if
           there is a shippable version.
         </p>
+        <p>
+          Currently, the full version is only available in German but we will
+          provide an English translation for the release version of Impromat.
+        </p>
         <h2>Offline Version</h2>
         <p>
           As long as you do not use the sign in option via Google you are using
           the offline version of Impromat. The offline version never
           synchronizes or transmitts any workshop creations or other data.
         </p>
+
+        <hr></hr>
+        <ReactMarkdown>{datenschutzMarkdown}</ReactMarkdown>
+        <a
+          href="https://datenschutz-generator.de/"
+          target="_blank"
+          rel="noreferrer"
+        >
+          Erstellt mit kostenlosem Datenschutz-Generator.de von Dr. Thomas
+          Schwenke
+        </a>
       </IonContent>
     </IonPage>
   );