+++ packages/impromat/src/pages/legal-markdown.ts
@@ -1,0 +1,8 @@
+export const legalMarkdown = `
+## Angaben gem. § 5 TMG:
+
+Marco Klein  
+Grünwalderstr. 41  
+81547 München  
+impromat@marcoklein.de
+`;